fix: DESKTOP - Remove the `DisableProfileImport` policy
This was added wayyyyy back at the beginning of Phoenix, mainly to remove Firefox's default "Import bookmarks" bookmark on the toolbar (to improve UX/declutter the UI).

But, our value for `browser.uiCustomization.state` appears to remove the bookmark anyways, independently of the policy, so it's not necessary for this. In general, I think this could be a useful feature for users, and I don't see any meaningful privacy or security risks with it, so, we can re-enable it (I think I probably shouldn't have added it in the first place TBH).

Signed-off-by: celenity <celenity@celenity.dev>