feat: Set the browser region to a dummy value ("XX") instead of "US"
This is preferable for various reasons. For instance, it ensures that Firefox Suggest and other unwanted functionality (ex. DoH rollout) are disabled by default. It also means that in the case of Firefox Suggest, we no longer have to lock the master pref for it to take effect - so it's now possible for Phoenix users to enable Suggest manually if desired.

In general, this also ensures that we aren't impacted by unwanted geo-targeting

Signed-off-by: celenity <celenity@celenity.dev>